Changed the locations of template lookups.
The 'src' directory is now searched before the 'local' or 'vendor'
directory to match the DI container's search priority.

The per-template file overrides has been changed to the full appName,
modName, and actName of the routed request instead of trying to
simplify the filename some and having lots of collissions.